Organize scripts by category
