+++ scripts/setup/setup_env.py
